further small clean ups
